Isolate executing query and handle errors
